+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.1] - 2019-07-12
 ### Added
-- Added `Struct` extensions
-- Added `Time.elapse`
+- Added String => `acronym`
+- Added Struct extensions
+- Added Time => `elapse`
 
 ## [1.0.0] - 2019-07-12
 ### Added